CLIP-DallE-SketchFab-001-main/requirements.txt
Flask==2.0.2
gunicorn
openai==0.27.8
torch
torchvision
openai-clip